refactor: replace guardian-ui Pill component with reusable StatusIndicator
